requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
paho-mqtt==1.6.1
python-nmap==0.7.1
requests==2.28.1